requirements.txt
asgiref==3.8.1
Django==5.1.2
django-ckeditor==6.7.1
django-etc==1.4.0
django-hitcount==1.3.5
django-jazzmin==3.0.1
django-js-asset==2.2.0
sqlparse==0.5.1
tzdata==2024.2
